+++ src/models/Settings.php
@@ -37,6 +37,12 @@ class Settings extends Model
     public bool $useNewBlockGrid = false;
 
     /**
+     * @var string|array|null The asset sources block type icons can be selected from, if using the new block grid.
+     * @since 3.6.0
+     */
+    public string|array|null $blockTypeIconSources = '*';
+
+    /**
      * @inheritdoc
      */
     protected function defineRules(): array